Ingest hbz Marc21 export and log errors to debug format issues
This commit also contains a few tweaks to make the project work
out of the box without checking out and building dependencies, and
some Eclipse project settings tweaks. This stuff is here only for
convenience (requiered morph config is present here), and will stay
in its branch or move to a different place in the future.